Use assertz when storing long term facts.
The long_term_fact should be a FIFO stack. It currently behaves as a
LIFO stack.

The long_term_fact stack is meant to store action that will be done
asynchronously. Let say I run

> botsito: dci sync OSP12
> botsito: dci sync OSP13

The bot would answer in the following way

** [DCI] Syncing of OSP13: failure
** [DCI] Syncing of OSP12: failure

OSP13 being dealt with first, when the user actually specified OSP12
first.